bring in updates from main repo

+++ files/scripts/syncd.sh
@@ -97,8 +97,13 @@ start() {
         fi
 
         if [[ x"$WARM_BOOT" != x"true" ]]; then
-            /bin/systemctl stop pmon
-            /usr/bin/hw-management.sh chipdown
+            if [[ x"$(/bin/systemctl is-active pmon)" == x"active" ]]; then
+                /bin/systemctl stop pmon
+                /usr/bin/hw-management.sh chipdown
+                /bin/systemctl restart pmon
+            else
+                /usr/bin/hw-management.sh chipdown
+            fi
         fi
 
         if [[ x"$BOOT_TYPE" == x"fast" ]]; then
@@ -108,10 +113,6 @@ start() {
         /usr/bin/mst start
         /usr/bin/mlnx-fw-upgrade.sh
         /etc/init.d/sxdkernel start
-
-        if [[ x"$WARM_BOOT" != x"true" ]]; then
-            /bin/systemctl start pmon
-        fi
     fi
 
     if [[ x"$WARM_BOOT" != x"true" ]]; then
@@ -123,10 +124,6 @@ start() {
     # start service docker
     /usr/bin/${SERVICE}.sh start $DEV
     debug "Started ${SERVICE} service..."
-
-    if [[ x"$sonic_asic_platform" == x"mellanox" && x"$BOOT_TYPE" == x"fast" ]]; then
-        /usr/bin/hw-management.sh chipupen
-    fi
 
     unlock_service_state_change
 }
